msm: camera: common: Fix compilation issues for 32-bit arch
Fix improper use of divide and modulo operator on 64 bit numbers.
Fix variable type to work with both 32/64 bit arch.

CRs-Fixed: 2543730

+++ drivers/cam_smmu/cam_smmu_api.c
@@ -3546,7 +3546,7 @@ static int cam_populate_smmu_context_banks(struct device *dev,
 	}
 
 	dma_set_max_seg_size(dev, DMA_BIT_MASK(32));
-	dma_set_seg_boundary(dev, DMA_BIT_MASK(64));
+	dma_set_seg_boundary(dev, (unsigned long)DMA_BIT_MASK(64));
 
 end:
 	/* increment count to next bank */
